Merge to latest Phantombot nightly

+++ source/me/mast3rplan/phantombot/twitchwsirc/Session.java
@@ -62,6 +62,8 @@ public class Session {
     private String botName;
     private String oAuth;
     private int chatLineCtr = 0;
+    private Long lastTry = 0L;
+    private Long lastModTry = 0L;
 
     /*
      * Creates an instance for a Session
@@ -91,7 +93,7 @@ public class Session {
      * @param  oAuth        OAUTH login
      * @param  Channel      Channel instance  
      * @param  eventBus     Eventbus
-     */ 
+     */
     private Session(Channel channel, String channelName, String botName, String oAuth, EventBus eventBus) {
         this.channelName = channelName.toLowerCase();
         this.eventBus = eventBus;
@@ -113,35 +115,28 @@ public class Session {
     }
 
     /*
-     * Attempts to reconnect to Twitch after a failure.
+     * Trys to reconnect to Twitch.
      */
     public void reconnect() {
         Boolean reconnected = false;
-        int     reconnectCtr = 0;
-        int     sleepTime = 5000;
-        int     maxAttempts = 50;
 
-        while (reconnectCtr != maxAttempts && !reconnected) {
-            reconnectCtr++;
-            try {
-                this.twitchWSIRC.delete();
-                this.twitchWSIRC = TwitchWSIRC.instance(new URI("wss://irc-ws.chat.twitch.tv"), channelName, botName, oAuth, channel, this, eventBus);
-            } catch (Exception ex) {
-                com.gmt2001.Console.err.println("TwitchWSIRC URI Failed, PhantomBot will exit: " + ex.getMessage());
-                System.exit(0);
-            }
-            reconnected = twitchWSIRC.connectWSS(true);
-            if (!reconnected) {
+        while (!reconnected) {
+            if (lastTry + 10000L <= System.currentTimeMillis()) {
+                lastTry = System.currentTimeMillis();
                 try {
-                    Thread.sleep(sleepTime);
-                } catch (InterruptedException ex) {
-                    com.gmt2001.Console.err.println("Sleep failed during reconnect: " + ex.getMessage());
+                    this.twitchWSIRC.delete();
+                    this.twitchWSIRC = TwitchWSIRC.instance(new URI("wss://irc-ws.chat.twitch.tv"), channelName, botName, oAuth, channel, this, eventBus);
+                    reconnected = this.twitchWSIRC.connectWSS(true);
+                } catch (Exception ex) {
+                    com.gmt2001.Console.err.println("Failed to reconnect to TwitchWSIRC... PhantomBot will now exit: " + ex.getMessage());
+                    System.exit(0);
                 }
             }
-        }
-        if (reconnectCtr == maxAttempts) {
-            com.gmt2001.Console.out.println("Unable to reconnect to Twitch WS-IRC after " + maxAttempts + " attempts, bot will shutdown.");
-            System.exit(0);
+            try {
+                Thread.sleep(1000);
+            } catch (InterruptedException ex) {
+                com.gmt2001.Console.debug.println("Sleep failed during reconnect: " + ex.getMessage());
+            }
         }
     }
 
@@ -223,7 +218,10 @@ public class Session {
      * Close the connection to WSIRC.
      */
     public void close() {
-        this.twitchWSIRC.close();
+        this.twitchWSIRC.delete();
+    }
+    public void close(String channelName) {
+        this.twitchWSIRC.delete(channelName);
     }
 
     /*
@@ -271,6 +269,19 @@ public class Session {
         if (message != null && sendMessages) { //make sure the message is not null, and make sure we are allowed to send messages.
             twitchWSIRC.send("PRIVMSG #" + channelName + " :" + message.message);// send the message to Twitch.
             com.gmt2001.Console.out.println("[CHAT] " + message.message);//print the message in the console once its sent to Twitch.
+        } else {
+            if (!sendMessages && message != null) {
+                com.gmt2001.Console.out.println();
+                com.gmt2001.Console.out.println("[ERROR] " + getNick() + " is not detected as a moderator!");
+                if (lastModTry + 60000 <= System.currentTimeMillis()) {
+                    lastModTry = System.currentTimeMillis();
+                    com.gmt2001.Console.out.println("[INFO] Trying to automatically detect moderator status now.");
+                    saySilent(".mods");
+                } else {
+                    com.gmt2001.Console.out.println("[INFO] Will try to automatically detect moderator status in 60 seconds.");
+                }
+                com.gmt2001.Console.out.println();
+            }
         }
     }
 
@@ -294,9 +305,22 @@ public class Session {
         }
 
         Message message = session.sendQueue.poll();
-        if (message != null && sendMessages) { 
-            twitchWSIRC.send("PRIVMSG #" + channelName + " :" + message.message);
-            com.gmt2001.Console.out.println("[CHAT] " + message.message);
+        if (message != null && sendMessages) { //make sure the message is not null, and make sure we are allowed to send messages.
+            twitchWSIRC.send("PRIVMSG #" + channelName + " :" + message.message);// send the message to Twitch.
+            com.gmt2001.Console.out.println("[CHAT] " + message.message);//print the message in the console once its sent to Twitch.
+        } else {
+            if (!sendMessages && message != null) {
+                com.gmt2001.Console.out.println();
+                com.gmt2001.Console.out.println("[ERROR] " + getNick() + " is not detected as a moderator!");
+                if (lastModTry + 60000 <= System.currentTimeMillis()) {
+                    lastModTry = System.currentTimeMillis();
+                    com.gmt2001.Console.out.println("[INFO] Trying to automatically detect moderator status now.");
+                    saySilent(".mods");
+                } else {
+                    com.gmt2001.Console.out.println("[INFO] Will try to automatically detect moderator status in 60 seconds.");
+                }
+                com.gmt2001.Console.out.println();
+            }
         }
     }
 
@@ -367,7 +391,7 @@ public class Session {
      
     class MessageTask extends TimerTask {
         private final Session session;
-        private final double messageLimit = PhantomBot.instance().getMessageInterval();
+        private final double messageLimit = PhantomBot.getMessageInterval();
         private long lastMessage = 0;
 
         public MessageTask(Session session) {
